+++ src/main/java/com/vomiter/rangedjs/projectile/ArrowHitEntityEventJS.java
@@ -1,9 +1,11 @@
 package com.vomiter.rangedjs.projectile;
 
+import net.minecraft.world.entity.projectile.AbstractArrow;
 import net.minecraft.world.entity.projectile.Projectile;
 import net.minecraft.world.phys.HitResult;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
+@SuppressWarnings("unused")
 public class ArrowHitEntityEventJS extends ProjectileHitEntityEventJS {
     private float damage;